+++ CHANGELOG.md
@@ -3,7 +3,13 @@ Changelog
 
 [Unreleased]
 ------------
--
+### Added
+- Auto inputting of room code from url's hash.
+  - Auto joining of room if default username is set.
+
+### Changed
+- Link with room code in hash is put in clipboard rather than just the room
+code on creating a room.
 
 [v1.1.1] - 2020-05-29
 ---------------------